requirement.txt
psycopg2-binary
flask
request
flask-sqlalchemy
Flask-Migrate
python-dotenv
gunicorn